requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
discord.py==2.3.2
docopt==0.6.2
frozenlist==1.4.1
idna==3.6
multidict==6.0.4
mutagen==1.47.0
pipreqs==0.4.13
pycparser==2.21
pycryptodomex==3.19.1
PyNaCl==1.5.0
requests==2.31.0
urllib3==2.1.0
websockets==12.0
yarg==0.1.9
yarl==1.9.4
yt-dlp==2023.12.30
